Added new powerups: Freeze, Freeze_Other, Plus50Points, Minus50Points

+++ src/org/academiadecodigo/twoballs/gameobjects/move/Speed.java
@@ -6,9 +6,9 @@ package org.academiadecodigo.twoballs.gameobjects.move;
 
 public class Speed {
 
-    public int x;
+    public float x;
 
-    public int y;
+    public float y;
 
     public Speed() {
